Changes for scheduling stopping dispatch command -- adding more precision by measuring from min to sec - added some logs for better error tracing

+++ hems-api-contracts/src/main/java/com/project/hems/hems_api_contracts/contract/dispatch/DeviceCommand.java
@@ -18,7 +18,7 @@ public class DeviceCommand {
     private UUID siteId; // The specific site
     private UUID programId;
     private DispatchMode mode;
-    private Integer durationMinutes;
+    private Long durationMinutes;
     private Long targetPowerW;
-    private Integer targetSoc;
+    private int targetSoc;
 }
